refactor(web): `getScheduleUseCase` 리팩토링
- 기존 일정과 출석 데이터를 한번에 호출하는 구조에서 일정과 출석을 각각 분리하여 호출하는 구조로 개선
