+++ apps/web/src/app/[space-slug]/_components/dashboard-attendance-section.tsx
@@ -1,6 +1,6 @@
 import { queryAllMembersUseCase } from "@/_pages/members/use-cases/query-all-members";
 import { DashboardAttendanceCard } from "@/_pages/schedule/ui/dashboard-attendance-card";
-import { getSchedulesUseCase } from "@/_pages/schedule/use-cases/get-schedules";
+import { getAttendanceStatusUseCase } from "@/_pages/schedule/use-cases/get-attendance-status";
 
 interface DashboardAttendanceSectionProps {
   spaceId: string;
@@ -9,10 +9,10 @@ interface DashboardAttendanceSectionProps {
 }
 
 export const DashboardAttendanceSection = async ({ spaceId, userId, slug }: DashboardAttendanceSectionProps) => {
-  const [scheduleData, members] = await Promise.all([
-    getSchedulesUseCase(spaceId, userId),
+  const [attendance, members] = await Promise.all([
+    getAttendanceStatusUseCase(spaceId, userId),
     queryAllMembersUseCase(spaceId),
   ]);
 
-  return <DashboardAttendanceCard slug={slug} attendance={scheduleData.attendance} totalMembers={members.length} />;
+  return <DashboardAttendanceCard slug={slug} attendance={attendance} totalMembers={members.length} />;
 };
