+++ apps/web/src/app/[space-slug]/schedule/_components/attendance-card-section.tsx
@@ -1,6 +1,6 @@
 import { queryAllMembersUseCase } from "@/_pages/members/use-cases/query-all-members";
 import { AttendanceCard } from "@/_pages/schedule";
-import { getSchedulesUseCase } from "@/_pages/schedule/use-cases/get-schedules";
+import { getAttendanceStatusUseCase } from "@/_pages/schedule/use-cases/get-attendance-status";
 import type { SpaceInfo } from "@/entities/spaces";
 import type { Requester } from "@/features/space/lib/assert-permission";
 
@@ -10,10 +10,10 @@ interface AttendanceCardSectionProps {
 }
 
 export const AttendanceCardSection = async ({ space, membership }: AttendanceCardSectionProps) => {
-  const [scheduleData, members] = await Promise.all([
-    getSchedulesUseCase(space.spaceId, membership.userId),
+  const [attendance, members] = await Promise.all([
+    getAttendanceStatusUseCase(space.spaceId, membership.userId),
     queryAllMembersUseCase(space.spaceId),
   ]);
 
-  return <AttendanceCard slug={space.slug} attendance={scheduleData.attendance} totalMembers={members.length} />;
+  return <AttendanceCard slug={space.slug} attendance={attendance} totalMembers={members.length} />;
 };
