+++ apps/web/src/app/[space-slug]/schedule/_components/schedule-section.tsx
@@ -1,5 +1,5 @@
 import { ScheduleList } from "@/_pages/schedule";
-import { getSchedulesUseCase } from "@/_pages/schedule/use-cases/get-schedules";
+import { getScheduleListUseCase } from "@/_pages/schedule/use-cases/get-schedule-list";
 import type { SpaceInfo } from "@/entities/spaces";
 import type { Requester } from "@/features/space/lib/assert-permission";
 
@@ -8,8 +8,8 @@ interface ScheduleSectionProps {
   membership: Requester;
 }
 
-export const ScheduleSection = async ({ space, membership }: ScheduleSectionProps) => {
-  const scheduleData = await getSchedulesUseCase(space.spaceId, membership.userId);
+export const ScheduleSection = async ({ space, membership: _membership }: ScheduleSectionProps) => {
+  const scheduleData = await getScheduleListUseCase(space.spaceId);
 
   return <ScheduleList slug={space.slug} upcoming={scheduleData.upcoming} expired={scheduleData.expired} />;
 };
